requirements.txt
# neuropulse-grant-agent — зависимости
# Установка: pip install -r requirements.txt

# HTTP-клиент (для API при необходимости)
requests>=2.28.0

# Переменные окружения
python-dotenv>=1.0.0

# Yandex AI Studio (REST API совместим с OpenAI-клиентом)
openai>=2.28.0

# Yandex AI Studio — Realtime API (голосовой агент, WebSocket)
aiohttp>=3.9.0
numpy>=1.26.0
sounddevice>=0.4.6

# Веб-чат для команды гранта
streamlit>=1.28.0

# Генерация документов Word по гранту
python-docx>=1.0.0

# Яндекс Календарь (CalDAV) для блока «Ближайшие сроки» в дашборде
caldav>=1.3.0

# Яндекс.Диск — выгрузка сгенерированных документов
yadisk>=0.6.0
